refactor(ui): split onboarding into kebab-case folder components
Each component now lives in its own kebab-case folder with an
index.tsx entry point. Props interfaces are extracted to colocated
types.ts files. The zod schema for the try-it-out form moves to
schema.ts so the derived TryItOutSubmission type can live in types.ts
without circular imports.
