+++ genai-engine/ui/src/components/onboarding/background-aurora/types.ts
@@ -1,0 +1,9 @@
+export interface AuroraBlob {
+  top: string;
+  left: string;
+  size: number;
+  color: string;
+  x: number[];
+  y: number[];
+  duration: number;
+}
