+++ genai-engine/ui/src/components/onboarding/landing-hero/types.ts
@@ -1,0 +1,4 @@
+export interface LandingHeroProps {
+  onTry: () => void;
+  onLogin: () => void;
+}
