+++ genai-engine/ui/src/components/onboarding/onboarding-options.ts
