+++ genai-engine/ui/src/components/onboarding/pick-card/types.ts
@@ -1,0 +1,11 @@
+import type { ReactNode } from "react";
+
+export interface PickCardProps {
+  variant: "primary" | "default";
+  icon: ReactNode;
+  badge?: string;
+  title: string;
+  description: string;
+  cta: string;
+  onClick: () => void;
+}
