+++ genai-engine/ui/src/components/onboarding/try-it-out-form/types.ts
@@ -1,0 +1,6 @@
+import type { TryItOutSubmission } from "./schema";
+
+export interface TryItOutFormProps {
+  onBack: () => void;
+  onSubmit: (data: TryItOutSubmission) => void;
+}
