+++ genai-engine/ui/src/components/onboarding/use-pointer-tracking.ts
